Added test for readonly-function
